requirements.txt
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.2.1
distlib==0.3.8
Faker==22.6.0
filelock==3.14.0
Flask==3.1.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
oauthlib==3.2.2
platformdirs==4.2.1
python-dateutil==2.8.2
python-dotenv==1.1.0
requests==2.31.0
requests-oauthlib==1.3.1
schedule==1.2.1
six==1.16.0
soupsieve==2.5
tweepy==4.14.0
urllib3==2.2.1
urlopen==1.0.0
virtualenv==20.26.1
Werkzeug==3.1.3
